+++ src/commands/index.ts
@@ -136,6 +136,7 @@ import { temporarycommandedit } from './moderation/temporaryCommandEdit';
 import { temporarycommand } from './general/temporaryCommand';
 import { yourControls } from './general/yourControls';
 import { notams } from './general/notams';
+import { vatsimData } from './utils/vatsimData';
 
 const commands: CommandDefinition[] = [
     ping,
@@ -274,6 +275,7 @@ const commands: CommandDefinition[] = [
     temporarycommand,
     yourControls,
     notams,
+    vatsimData,
 ];
 
 const commandsObject: { [k: string]: CommandDefinition } = {};